CQ-4242735 - Make PageModelManager context path aware (#14)
The original bug observed is that the navigation component isn't displayed on a server that starts with a context path. This is because the navigation components tries to fetch the model from the server without considering the context path, and thus, receives no model data for the navigation.

To prevent this, we need to extend the PageModelManager module to make it context path aware. To avoid introducing a Granite dependency in this module (cf.Granite.HTTP.detectContextPath()), the idea is to make the PageModelManager detect the context path based on the current location; when it initializes. Afterwards, all subsequent GET calls to the server would be externalized by prepending the context path to the requested URL.